PR 25540: Add CachedDataSource and unit tests
 - Add CachedDataSource and unit tests

It is an optional (opt-in) enhanced version of the base DataSource that comes with an automatic, short-lived cache.

The unit tests describe expected functionality.

The primary motivation for this is for us in odsp-next to communicate with datasources in sphome (which uses their own set of datasources, look up oil.ds.ts) through a common cache. It thus implements various capabilities of that including automatic partitioning.