PR 25638: Remove inconsistent stripping of privates from json-stringify if it points to value
Line 118-121 of JSON.stringify in BaseDataStore.setValue below discards private objects from being written into stringified and written into cache, IF the value reference by the private key is an object.

That makes no sense to me – if we’re discarding privates, we should discard all privates (including privates to strings, privates to numbers).

I thus suspect that these 3 lines are an unintended bug, and am going to remove it. Given that privates of all other data types are being written to cache, I can only assume that private stripping is not an actual necessary scenario (and also the lack of stripping is not affecting anyone, given how long this inconsistency has been here).