PR 30019: Pick up new buildtools version with new lint rules
Pick up buildtools version 36, which no longer includes TS 2 changes but does include new lint rules. This is to prevent people from checking in things that violate the new lint rules before the TS 2 change is ready, to make the change easier.

(Also includes a lingering test that didn't get moved from odsp-shared with the rest of the string utilities...)