Merged PR 37431: Add unit tests for follow button
Added 5 new tests to cover the follow feature in site header. This has almost 100% code coverage for all follow code, the only thing it does not test is cache interaction and error handling if call to spsocial api fails (but the behavior for error case is not specified and not required to be locked down/tested).

The process of writing tests actually uncovered a subtle bug :\