Merged PR 38225: Group membership: display Guest info, check for malformed members list
Two changes here:

(1) Use the UserType attribute to determine whether a member is a guest. Mark guests as "Guest" and do not show the dropdown menu because we don't want to allow promoting guests to owners or removing guests (until the ability to add them via SharePoint is implemented).

(2) In theory, the members list received from the server is supposed to contain all members, including owners. However, this is not enforced by the server and there seems to be an endpoint somewhere that is adding owners to the owners list without adding them to the members list. Therefore, I rewrote some code to make sure we are displaying all owners, even owners who are missing from the members list.

Related work items: #290563