Merged PR 39571: Remove panel margin workaround
We originally introduced this margin workaround so that our panels would clear the suitenav area at the top of the page. This was due to a z-index override set on the suite nav. Now the suitenav has pushed their updated UI to first release and soon 100% of prod and the z-index issues are gone, we can remove this workaround.