Merged PR 39642: Hide "Edit (link to edit file)" column
This is a very very old column type (from back before SP had context menus) that in classic would display a pencil icon and go to the edit form when clicked. Currently the column is displayed as blank, which isn't good, and design decided that since you can do the same action through the command bar and context menu this column should be hidden.

Related work items: #252605