Merged PR 40571: Clear cache after adding user/users to group during group creation.
Clear cache after add user/users to group during group creation.
Because create group is using principle name, while clear cache is using user id, so the cache is not cleared as it supposes to when members added to group during group creation. In the fix, if the user id is not exist, it will be obtained from getCurrentUser.

Here are the deploy links:
Groups (EDOG):
https://msft.spoppe.com/teams/SPGroups/playground/cookieredirect/redirect.aspx?mobile=0&setcookie=https%3A%2F%2Fresourceseng.blob.core.windows.net%2Ffiles%2Fdev-zeqiu-zeqiu-pc02c--git-odsp-next-app-min-master%2F&redirect=https%3A%2F%2Fmsft.spoppe.com%2Fsites%2Focw%2F
Groups (MSIT):
https://microsoft.sharepoint.com/teams/cyrusplayground/cookieredirect/redirect.aspx?mobile=0&setCookie=https%3A%2F%2Fresourceseng.blob.core.windows.net%2Ffiles%2Fdev-zeqiu-zeqiu-pc02c--git-odsp-next-app-min-master%2F&redirect=https%3A%2F%2Fmicrosoft.sharepoint.com%2Fteams%2Fsts-dev%2F_layouts%2F15%2FGroup.aspx%3FGroupId%3Dd80a128e%252Dfc33%252D495a%252D8f5b%252Da4fda082f864%26AppId%3DFiles

Related work items: #290585