Merged PR 48941: Custom Field Formatter module
This is the initial checkin for a no-code custom field renderer. The core logic is in CustomFormatter.ts, with interfaces defined in CustomFormatter.Props.ts.
Right now, nobody uses these classes, except the example page, so there should be no bundle size changes or anything like that. It should essentially be a no-op from a bundle perspective.... @<rchen@microsoft.com> , and @<danst@microsoft.com> , can you please make sure that indeed this will not cause any bundles to include this module yet... I think it's pretty obvious, but just want to make sure..