Merged PR 50862: Move column creation errors into the panel.
Move column creation errors into the create column panel, so that people don't lose progress when they hit errors.
Rename the ICreateFieldOptions to IFieldSchema, given that it is going to be used a lot more in the future for edit as well, and move it to the top level interfaces folder.
Rename CreateColumnPanel to ColumnManagementPanel to reflect its role as both create and edit panel.
Move IServerField interface to odsp-datasources.