Merged PR 51770: Merge zhiliu/fixForOnePage to master
I 'd like to use loadModule utility in sp-client which is not require js based. The scenario is one page navigation from modern home page to doc lib. I will load require js on-demand. but if I import loadModule component, this RequireJSErrorHandler will also be included and will start overloading onError of requireJS which is a null reference exception.  Make the logic to overriding require js onError to be a function, so other code can call it on-demand.