Merged PR 57444: Fix update of member count in site header
The bug: When the group membership panel is closed, it reloads the group members to trigger an update of the facepile and member count in the site header. The load method I was using sometimes ends up hitting the cache, and the member count does not update.

Solution: Add an option to force bypassing the cache. I also added some logic to only reload the members if the user added/removed members while using the panel.

Note: The Membership code has become pretty complicated, with lots of logic about whether to use the cache or not.  When the VirtualGroupMembersList feature is done rolling out, I hope to be able to simplify this code (removing the allMembers option for example).  For loading large numbers of members, it makes more sense to use the new MembershipPager.