Merged PR 58976: Fix race condition
This was an interesting intermittent error: when the user scrolled to the bottom of the members list, sometimes the next page wouldn't be requested from the server.  It turned out that sometimes the MembersList component would re-render before the isPageLoading flag was reset to false.  As a result, when the spinner at the bottom of the MembersList re-rendered (supposed to trigger a request for the next page), it would think the page load was already in progress and do nothing. The fix is simply to set isPageLoading to false before calling setState.