Merged PR 61383: Merge twu/ExcelSurveyV2FormPropertyInListContext to master
Add canUserCreateMicrosoftForm property ISPGetItemResponse.

We are adding the canUserCreatMicrosoftForm property in list context. The page context in personal site is constant. So that we need to check if Microsoft Forms can be enabled when user navigate to the team site document library from their personal site. The canUserCreatMicrosoftForm property will be consumed from IListFacet.ts.