Merged PR 63487: Rename useFiltersInViewXml in getItemContext to ignoreFilterParams to make it more clear. And modify the getUrl function in SPListItemRetriever to take whole context as input param
GetUrl now need to look at ignoreFilterParams in the getItemContext to determine whether to add filter param in the request Url. By ignoreFilterParams is only set for smartfilters scenario.

The reason is that we cannot correctly determine whether to add filter params in the request Url based on existing properties in getItemContext. We need to have specific flag for it to ensure we don't break any existing scenarios depending on filter params in the request url.