Merged PR 65408: Enable localization string export in web-library-build projects
This change enables building RESX strings in odsp-shared-react and other web-library-build projects. RESX is exported twice for sp-client projects (lib directory) and odsp-next (lib-amd). This saves us from needing to duplicate strings but you still can't consume the strings directly in odsp-common code because of loader incompatibilities between webpack and the onedrive-buildtools bundler. Will resolve this in the future.

NOTE: Not ready for use yet, still need a further change to hookup publishing and consuming LCL and LCG from loc team, you will only get en-us and pseudo locales for now. I'll send mail with further details when it's ready.

Also moving web-library-build projects to use an internal shared build rig (@ms/odsp-build) that adds in the localization tasks and other configuration that is common between packages in this repo.