Merged PR 72009: Hide Overflow Button from Share Dialog when Manage Access Panel is Already Open
With new change we made in manage access panel, user can access the share dialog via the Share commandbar item. If user clicks on the overflow button in the share dialog and hit "Manage Access", another manage access panel opens on top of the currently opened one.

Adding a Prop "hideShareDialogOverflowButton" in IShareProps, and if the prop is true we don't show the overflow button in the share dialog.

**OneDrive for Business (MSIT):**
https://microsoft.sharepoint.com/teams/cyrusplayground/cookieredirect/redirect.aspx?mobile=0&setCookie=https%3A%2F%2Fresourceseng.blob.core.windows.net%2Ffiles%2Fdev-leji-desktop-l4jb9g1c--git-odsp-next-app-min-leji%2Fupdate-aria-label-for-shared-glyph%2F